Fix DataLoader iter(dataloader) cannot be called more than once (#544)
PyTorch DataLoader allows iter(datalaoder) to be called more than once. Specifically, the user can define a function:
```
def train_one_epoch(dataloader, ...):
    for batch in dataloader:
        ...
```
and use the function with the same dataloader:
```
with DataLoader(...) as dataloader:
    for epoch in range(num_epochs):
        train_one_epoch(dataloader, ...)
```
In the current implementation, when `for batch in dataloader` is called in the second epoch, no batch will be returned.

This PR will enable the usage above.

### Fix
When each new iteration starts: reset iterator internal status: including
* create new _shuffling_buffer
* reset petastorm reader

And add a limit, only after we finish the last iteration, we can start a new iteration because these iterators share the same underlying reader, so they cannot run in parallel.